WIP: Add SiliconLabs/arduino framework and Sparkfun Thing Plus Matter board
This adds some as some initial boilerplate code for a platform builder script for
https://github.com/SiliconLabs/arduino .
That's the official Arduino core for the upcoming Arduino Nano Matter board.

It also adds a board definition for the Sparkfun Thing Plus Matter board.
I'm fairly sure I've deduced the correct MCU data for this board from its
spec sheet and some other pointers but I can't guarantee
it is _actually_ an EFR32MG24B020F1536IM40.

The svd file for the MCU was sourced from
https://www.keil.arm.com/family/silicon-labs-efr32mg24-series/